Move enumerations into a separate module container_attributes
Having the enums in a separate module allows us to import it in other modules
without having import cycles
