+++ src/bci_build/package/apache_tomcat.py
@@ -2,6 +2,7 @@
 
 import datetime
 
+from bci_build.container_attributes import PackageType
 from bci_build.containercrate import ContainerCrate
 from bci_build.os_version import CAN_BE_LATEST_OS_VERSION
 from bci_build.os_version import OsVersion
@@ -9,7 +10,6 @@ from bci_build.package import DOCKERFILE_RUN
 from bci_build.package import ApplicationStackContainer
 from bci_build.package import OsContainer
 from bci_build.package import Package
-from bci_build.package import PackageType
 from bci_build.package import Replacement
 from bci_build.package import _build_tag_prefix
 from bci_build.registry import publish_registry
