+++ src/bci_build/package/appcontainers.py
@@ -2,18 +2,18 @@
 
 from pathlib import Path
 
+from bci_build.container_attributes import BuildType
+from bci_build.container_attributes import PackageType
+from bci_build.container_attributes import SupportLevel
 from bci_build.os_version import ALL_NONBASE_OS_VERSIONS
 from bci_build.os_version import CAN_BE_LATEST_OS_VERSION
 from bci_build.os_version import OsVersion
 from bci_build.package import DOCKERFILE_RUN
 from bci_build.package import ApplicationStackContainer
-from bci_build.package import BuildType
 from bci_build.package import OsContainer
 from bci_build.package import Package
-from bci_build.package import PackageType
 from bci_build.package import ParseVersion
 from bci_build.package import Replacement
-from bci_build.package import SupportLevel
 from bci_build.package import _build_tag_prefix
 from bci_build.package.helpers import generate_from_image_tag
 from bci_build.package.helpers import generate_package_version_check
