+++ src/bci_build/package/base.py
@@ -5,13 +5,13 @@ from pathlib import Path
 
 from jinja2 import Template
 
+from bci_build.container_attributes import Arch
+from bci_build.container_attributes import BuildType
+from bci_build.container_attributes import PackageType
+from bci_build.container_attributes import SupportLevel
 from bci_build.os_version import OsVersion
-from bci_build.package import Arch
-from bci_build.package import BuildType
 from bci_build.package import OsContainer
 from bci_build.package import Package
-from bci_build.package import PackageType
-from bci_build.package import SupportLevel
 
 
 def _get_base_config_sh_script(os_version: OsVersion) -> str:
