+++ src/bci_build/package/golang.py
@@ -4,13 +4,13 @@ import textwrap
 from itertools import product
 from typing import Literal
 
+from bci_build.container_attributes import SupportLevel
 from bci_build.os_version import CAN_BE_LATEST_OS_VERSION
 from bci_build.os_version import OsVersion
 from bci_build.package import DOCKERFILE_RUN
 from bci_build.package import LOG_CLEAN
 from bci_build.package import DevelopmentContainer
 from bci_build.package import Replacement
-from bci_build.package import SupportLevel
 from bci_build.package import generate_disk_size_constraints
 
 _GO_VER_T = Literal["1.20", "1.21", "1.22", "1.23"]
