+++ src/bci_build/package/kiwi.py
@@ -1,9 +1,9 @@
 """KIWI Appliances Builder SDK container for easy appliance building on SLE Micro."""
 
+from bci_build.container_attributes import BuildType
 from bci_build.os_version import ALL_NONBASE_OS_VERSIONS
 from bci_build.os_version import CAN_BE_LATEST_OS_VERSION
 from bci_build.os_version import OsVersion
-from bci_build.package import BuildType
 from bci_build.package import DevelopmentContainer
 from bci_build.package import ParseVersion
 from bci_build.package import generate_disk_size_constraints
