+++ src/bci_build/package/node.py
@@ -3,12 +3,12 @@
 import datetime
 from typing import Literal
 
+from bci_build.container_attributes import SupportLevel
 from bci_build.os_version import CAN_BE_LATEST_OS_VERSION
 from bci_build.os_version import _SUPPORTED_UNTIL_SLE
 from bci_build.os_version import OsVersion
 from bci_build.package import DevelopmentContainer
 from bci_build.package import Replacement
-from bci_build.package import SupportLevel
 
 _NODE_VERSIONS = Literal[16, 18, 20, 21, 22, 23, 24]
