+++ src/bci_build/package/postfix.py
@@ -2,6 +2,7 @@
 
 from pathlib import Path
 
+from bci_build.container_attributes import SupportLevel
 from bci_build.os_version import ALL_NONBASE_OS_VERSIONS
 from bci_build.os_version import CAN_BE_LATEST_OS_VERSION
 from bci_build.os_version import OsVersion
@@ -9,7 +10,6 @@ from bci_build.package import DOCKERFILE_RUN
 from bci_build.package import ApplicationStackContainer
 from bci_build.package import ParseVersion
 from bci_build.package import Replacement
-from bci_build.package import SupportLevel
 from bci_build.package.helpers import generate_from_image_tag
 from bci_build.registry import ApplicationCollectionRegistry
