+++ src/bci_build/package/rust.py
@@ -3,11 +3,11 @@
 import datetime
 from itertools import product
 
+from bci_build.container_attributes import SupportLevel
 from bci_build.os_version import ALL_NONBASE_OS_VERSIONS
 from bci_build.os_version import CAN_BE_LATEST_OS_VERSION
 from bci_build.package import DevelopmentContainer
 from bci_build.package import Replacement
-from bci_build.package import SupportLevel
 from bci_build.package import generate_disk_size_constraints
 
 _RUST_GCC_PATH = "/usr/local/bin/gcc"
