+++ src/bci_build/package/spack.py
@@ -1,13 +1,13 @@
 """HPC container for Spack package manager."""
 
+from bci_build.container_attributes import Arch
+from bci_build.container_attributes import SupportLevel
 from bci_build.os_version import CAN_BE_LATEST_OS_VERSION
 from bci_build.os_version import _SUPPORTED_UNTIL_SLE
 from bci_build.os_version import OsVersion
 from bci_build.package import DOCKERFILE_RUN
-from bci_build.package import Arch
 from bci_build.package import DevelopmentContainer
 from bci_build.package import ParseVersion
-from bci_build.package import SupportLevel
 from bci_build.package import generate_disk_size_constraints
 from bci_build.package.helpers import generate_package_version_check
 from bci_build.package.versions import format_version
